pyproject.toml
[project]
name = "fifteenpuzzlebot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot[all]>=22.5",
]
